backend-20: * rewrote mail transport as a secondary adapter and added mail repository
