+++ backend/src/shared/types/container-type-keys.ts
@@ -4,8 +4,10 @@ const CONTAINER_TYPES = {
   UserService: Symbol.for('UserService'),
   MailService: Symbol.for('MailService'),
   UserRepository: Symbol.for('UserRepository'),
+  MailRepository: Symbol.for('MailRepository'),
 
   PrismaClient: Symbol.for('PrismaClient'),
+  MailTransporter: Symbol.for('MailTransporter'),
 };
 
 export { CONTAINER_TYPES };
